+++ GFF2019_Team4 Conveni/Assets/Scripts/BigSpiderController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 97e9b7aa030a6354ca873c315118aa8d
+guid: 0eade69ce3af7bf42b4aaed7271073b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
