--- GFF2019_Team4 Conveni/Assets/Scripts/Sensing.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0eade69ce3af7bf42b4aaed7271073b1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
